Update facebook sample
