+++ toro-sample/src/main/java/im/ene/lab/toro/sample/presentation/facebook/FbItemViewHolder.java
@@ -16,15 +16,25 @@
 
 package im.ene.lab.toro.sample.presentation.facebook;
 
+import android.animation.Animator;
+import android.animation.AnimatorListenerAdapter;
+import android.net.Uri;
 import android.support.annotation.IntDef;
 import android.support.annotation.Nullable;
 import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ImageView;
+import android.widget.TextView;
+import com.squareup.picasso.Picasso;
+import im.ene.lab.toro.ext.ToroVideoViewHolder;
+import im.ene.lab.toro.media.Cineer;
+import im.ene.lab.toro.media.LastMomentCallback;
+import im.ene.lab.toro.media.PlaybackException;
+import im.ene.lab.toro.player.widget.VideoPlayerView;
 import im.ene.lab.toro.sample.R;
-import im.ene.lab.toro.sample.develop.viewholder.SimpleToroVideoViewHolder;
+import im.ene.lab.toro.sample.data.SimpleVideoObject;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
@@ -107,22 +117,136 @@ public abstract class FbItemViewHolder extends RecyclerView.ViewHolder {
     }
   }
 
-  static class VideoPost extends SimpleToroVideoViewHolder {
+  static class VideoPost extends ToroVideoViewHolder implements LastMomentCallback {
 
     static final int LAYOUT_RES = R.layout.vh_fb_feed_post_video;
 
+    private ImageView mThumbnail;
+    private TextView mInfo;
+    private boolean isPlayable = false;
+    private boolean isReleased = false;
+    private long latestPosition = 0;
+
     public VideoPost(View itemView) {
       super(itemView);
+      mThumbnail = (ImageView) itemView.findViewById(R.id.thumbnail);
+      mInfo = (TextView) itemView.findViewById(R.id.info);
+      mVideoView.setLastMomentCallback(this);
+    }
+
+    @Override protected VideoPlayerView findVideoView(View itemView) {
+      return (VideoPlayerView) itemView.findViewById(R.id.video);
+    }
+
+    @Override public void setOnItemClickListener(View.OnClickListener listener) {
+      super.setOnItemClickListener(listener);
+      mInfo.setOnClickListener(listener);
+    }
+
+    private SimpleVideoObject mItem;
+
+    @Override public void bind(Object item) {
+      if (!(item instanceof SimpleVideoObject)) {
+        throw new IllegalStateException("Unexpected object: " + item.toString());
+      }
+
+      mItem = (SimpleVideoObject) item;
+      mVideoView.setMedia(Uri.parse(mItem.video));
+    }
+
+    @Override public boolean wantsToPlay() {
+      return isPlayable && visibleAreaOffset() >= 0.75;
+    }
+
+    @Nullable @Override public String getVideoId() {
+      return mItem.toString() + "@" + getAdapterPosition();
+    }
+
+    @Override public void onVideoPreparing() {
+      super.onVideoPreparing();
+      mInfo.setText("Preparing");
+    }
+
+    @Override public void onVideoPrepared(Cineer mp) {
+      super.onVideoPrepared(mp);
+      isPlayable = true;
+      mInfo.setText("Prepared");
+      latestPosition = 0;
+      isReleased = false;
+    }
+
+    @Override public void onViewHolderBound() {
+      super.onViewHolderBound();
+      Picasso.with(itemView.getContext())
+          .load(R.drawable.toro_place_holder)
+          .fit()
+          .centerInside()
+          .into(mThumbnail);
+      mInfo.setText("Bound");
+    }
+
+    @Override public void onPlaybackStarted() {
+      mThumbnail.animate().alpha(0.f).setDuration(250).setListener(new AnimatorListenerAdapter() {
+        @Override public void onAnimationEnd(Animator animation) {
+          VideoPost.super.onPlaybackStarted();
+        }
+      }).start();
+      mInfo.setText("Started");
+    }
+
+    @Override public void onPlaybackPaused() {
+      mThumbnail.animate().alpha(1.f).setDuration(250).setListener(new AnimatorListenerAdapter() {
+        @Override public void onAnimationEnd(Animator animation) {
+          VideoPost.super.onPlaybackPaused();
+        }
+      }).start();
+      mInfo.setText("Paused");
+    }
+
+    @Override public void onPlaybackCompleted() {
+      isPlayable = false;
+      mThumbnail.animate().alpha(1.f).setDuration(250).setListener(new AnimatorListenerAdapter() {
+        @Override public void onAnimationEnd(Animator animation) {
+          VideoPost.super.onPlaybackCompleted();
+        }
+      }).start();
+      mInfo.setText("Completed");
+    }
+
+    @Override public boolean onPlaybackError(Cineer mp, PlaybackException error) {
+      isPlayable = false;
+      mThumbnail.animate().alpha(1.f).setDuration(250).setListener(new AnimatorListenerAdapter() {
+        @Override public void onAnimationEnd(Animator animation) {
+          VideoPost.super.onPlaybackCompleted();
+        }
+      }).start();
+      mInfo.setText("Error: videoId = " + getVideoId());
+      return super.onPlaybackError(mp, error);
+    }
+
+    @Override public boolean isLoopAble() {
+      return true;
+    }
+
+    @Override public String toString() {
+      return "Video: " + getVideoId();
+    }
+
+    @Override public long getCurrentPosition() {
+      if (!isReleased) {
+        latestPosition = super.getCurrentPosition();
+      }
+
+      return latestPosition;
+    }
+
+    @Override public void onLastMoment(Cineer player) {
+      isReleased = true;
+      latestPosition = player.getCurrentPosition();
     }
 
     @Override protected boolean allowLongPressSupport() {
       return true;
     }
-
-    @Override public void bind(Object item) {
-      super.bind(item);
-      Log.d("VideoPlayerManager@VH", "pos = " + getAdapterPosition() + " | item = [" + item + "]");
-    }
-
   }
 }
